requirements.txt
PyGithub==2.1.1
nltk==3.8.1
wordcloud==1.9.3
matplotlib==3.8.3 